Add schema to confest instead of fixtures
